requirements.txt
flask==3.0.2
mistralai==0.0.12
numpy==1.26.4
python-dotenv==1.0.1
flask-cors==4.0.0
gunicorn==21.2.0 